Add files via upload
Nj parcels webscrapers built to collect data from 500 city in New Jersey. The crawler is capable of collecting data from 274 city and sending to a spread sheet on the google servers. It uses requests as the main library to pull for the website https://www.njparcels.com. The application uses threading to scrap multiple documents at the same time to reduces the time needed to complete the scraping. The code needs to be upgraded to using selenium as the main library with threading, or requests has to be configured to scrape all 500 pages o data. The code is under maintainance